Changed to new Dart AST
