+++ .gitignore
@@ -5,6 +5,7 @@ obj/
 .packages
 *.lock
 *.ast.json
+model.json
 **/*.Android/**/[Rr]esource.[Dd]esigner.cs
 **/*.Droid/**/[Rr]esource.[Dd]esigner.cs
 **/Android/**/[Rr]esource.[Dd]esigner.cs
