--- Dart2CSharpTranspiler/CSharp/Writer.cs
@@ -1,10 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Dart2CSharpTranspiler.CSharp
-{
-    public static class Writer
-    {
-    }
-}
